more issue fixed.

+++ app/src/main/java/amir/app/business/models/Location.java
@@ -3,6 +3,6 @@ package amir.app.business.models;
 import com.strongloop.android.loopback.Model;
 
 public class Location extends Model {
-    public int lat;
-    public int lng;
+    public float lat;
+    public float lng;
 }
